ent-challenges/enterprise-contract-agent/requirements.txt
# Core Dependencies
langgraph==0.2.28
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.0

# PDF Processing
pypdf==4.3.1
pymupdf==1.24.10

# Configuration
python-dotenv==1.0.1
pydantic==2.9.2

# Observability
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-exporter-otlp==1.27.0
prometheus-client==0.21.0

# Security - PII Detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.5

# PDF Generation (for sample contracts)
reportlab==4.2.5

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Utilities
requests==2.32.3
